+++ packages/tooltip/src/components/popover/ActionPrompt.tsx
@@ -1,0 +1,48 @@
+import { Column, Row, Space, StandardText } from "@stenajs-webui/core";
+import { FlatButton, Progress, StandardButton } from "@stenajs-webui/elements";
+import * as React from "react";
+
+interface ActionPromptProps {
+  text?: string;
+  yesLabel?: string;
+  noLabel?: string;
+  onYes?: () => void;
+  onNo?: () => void;
+  loading?: boolean;
+}
+
+export const ActionPrompt: React.FC<ActionPromptProps> = ({
+  text = "Are you sure?",
+  yesLabel = "Yes",
+  noLabel = "No",
+  onNo,
+  onYes,
+  loading
+}) => {
+  return (
+    <Column position={"relative"}>
+      <div style={{ opacity: loading ? 0 : 1 }}>
+        <StandardText>{text}</StandardText>
+        <Space />
+        <Row justifyContent={"flex-end"} alignItems={"center"}>
+          <FlatButton label={noLabel} onClick={onNo} />
+          <Space />
+          <StandardButton label={yesLabel} onClick={onYes} />
+        </Row>
+      </div>
+      {loading && (
+        <Row
+          justifyContent={"center"}
+          alignItems={"center"}
+          left={0}
+          top={0}
+          width={"100%"}
+          height={"100%"}
+          position={"absolute"}
+        >
+          <Progress size={"small"} />
+        </Row>
+      )}
+    </Column>
+  );
+};